refactor(copyright): extract tree_walk sub-module from detector.rs
Move parse-tree extraction functions (3195 lines) including
extract_from_tree_nodes, extract_from_spans, extract_bare_copyrights,
and ~30 helper predicates into detector_tree_walk.rs.
